+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - :sparkles: 完善 `HexUtils` 工具类。
 - :sparkles: 添加 JsonUtil，支持 `Jackson`, `Fastjson1`, `Fastjson2`, `Gson`, `hutool-json`。
 - :sparkles: copy 并采用时间轮优化 `TimedCache`，用于处理 mqtt 保留 session。
+- :sparkles: ChannelContext 规范化，`userId` 改为 `private`。 
 - :arrow_up: 依赖升级
 
 ### v0.0.8 - 2023-03-05