+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 发行版本
 
-### v0.0.9 - 2023-03-26
+### v0.0.9 - 2023-03-23
 - :fire: 优化 `ByteBufferUtil` 工具类，只 `dump` 有效的 `ByteBuffer` 数据和优化 `toString` 方法。
 - :sparkles: `HttpRequestHandler#clearStaticResCache` 方法改为 default 方法。
 - :sparkles: 支持空 `websocket` 包